Move lib methods to lib, separate keys to avoid circular dep
